requirements.txt
asgiref==3.6.0
Django==4.1.6
djangorestframework==3.14.0
Faker==16.6.1
Pillow==9.4.0
python-dateutil==2.8.2